apps/auth/requirements.txt
fastapi
uvicorn
pyjwt